Implement conditional responses to mock client
Following discussions at https://github.com/php-http/mock-client/issues/3 this
adds a method to set up the MockClient with responses or exceptions that are
used when a request matcher matches the request. Additionally, a callback can
be given to create responses / exceptions dynamically.

This allows for mock clients to not depend on the order the responses/exceptions
are mocked.

It also allows for responses to depend on some content in the request, avoiding
duplication in some cases.