Add conditional responses to mock client
Following discussions at https://github.com/php-http/mock-client/issues/3 this adds
a on method taking a RequestMatcher and then either a Response, an Exception or a
callable.

This allows for mock clients to not depend on the order the responses/exceptions
are mocked.

It also allows for responses to depend on some content in the request, avoiding
duplication in some cases.